+++ src/components/modal/index.tsx
@@ -1,7 +1,6 @@
 import DeleteAccountModal from './DeleteAccountModal';
 import DeleteProjectModal from './DeleteProjectModal';
 import EmailUpdateModal from './EmailUpdateModal';
-import ModifyProjectModal from './ModifyProjectModal';
 import PasswordUpdateModal from './PasswordUpdateModal';
 import ProfileUpdateModel from './ProfileUpdateModal';
 import WithdrawProjectModal from './WithdrawProjectModal';
@@ -10,7 +9,6 @@ export {
   DeleteAccountModal,
   DeleteProjectModal,
   EmailUpdateModal,
-  ModifyProjectModal,
   PasswordUpdateModal,
   ProfileUpdateModel,
   WithdrawProjectModal,
