+++ src/pages/projects/works/ProjectWorkBoards.tsx
@@ -1,9 +1,9 @@
 // import Add from '@assets/add.svg';
 import { Button } from '@components/common/Button';
-import CreateProjectModal from '@components/modal/CreateProjectModal';
-import { useModal } from '@hooks';
-import styled from 'styled-components';
 import ProjectWorkBoardItem from '@components/project/ProjectWorkBoardItem';
+import { useModal } from '@hooks';
+import CreateProjectModal from '@pages/projects/ProjectBoards/CreateProjectModal/CreateProjectModal';
+import styled from 'styled-components';
 
 const Section = styled.section`
   display: flex;
@@ -46,8 +46,6 @@ const ProjectWorkList = styled.ul`
 //   }
 // `;
 
-
-
 export interface Project {
   projectId: number;
   title: string;
@@ -63,7 +61,6 @@ const ProjectWorkBoards = () => {
   //   useModal();
 
   const [openModal] = useModal();
-
 
   return (
     <Section>
@@ -86,7 +83,7 @@ const ProjectWorkBoards = () => {
         </ProjectAddButton> */}
       </ProjectBoardHeader>
       <ProjectWorkList>
-        <ProjectWorkBoardItem/>
+        <ProjectWorkBoardItem />
       </ProjectWorkList>
     </Section>
   );
